backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi==0.111.0",
    "psycopg2-binary==2.9.9",
    "pydantic-settings==2.1.0",
    "pyjwt==2.8.0",
    "python-dotenv==1.0.1",
    "python-multipart==0.0.9",
    "qdrant-client==1.11.1",
    "sentence-transformers==3.0.1",
    "sqlmodel==0.0.15",
    "uvicorn[standard]==0.30.1",
]
